diverging startup
It seems that I have been able to arrive at a satisfactory solution by creating a global variable `mode` on shiny app startup. This allows me to choose if `app_ui` should be wrapped by `secure_app` in `run_app` function. This means that pool is established as a global, whereas for the local version it must be established later.